requirements.txt
pandas
PyMuPDF
resend
pathlib
dotenv
requests
zai-sdk
python-docx
chromadb
asyncio
fastapi[standard]
toon-llm
datetime
uuid
redis